Add files via upload
The creation of the "idx_hotel_id" bitmap index on the HOTEL_ID column in the BOOKINGS table aims to boost query speed. This type of index is useful when there are a few unique hotel IDs. It works like a map, associating each ID with a set of bits. These speeds up searches and filters related to hotels in the BOOKINGS table because the database can quickly check the bits instead of going through each row individually. In simpler terms, it's like having a cheat sheet that makes finding and using hotel information much faster