Add files via upload
The bar graph that is displayed shows the average ratings that different hotels received between 2015 and 2022. The Alpine Club is notable for having the highest average rating, outperforming a number of other well-known hotels. This information highlights how well it performed in sustaining high levels of customer satisfaction over the given time frame.